refactor: Rewrite how <Board /> is rendered
I've rewritten how <Board /> is rendered to make its size easily created
rather than being hard coded, which results to a harder way specifying
the size. This is done by using 2 loops--one for rows and the other for
squares per row.

Now, to modify the size of the <Board />, just give the numbers of rows
and squares per row rather than hard coding the JSX.